se agrega alerta de confirmacion al eliminar

+++ src/app/pages/clients/list-client/list-client.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { Cliente } from 'src/app/interfaces/client.interface';
 import { ClientService } from 'src/app/services/client.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-client',
@@ -41,16 +42,27 @@ export class ListClientComponent {
 
   deleteCliente(idCliente: any) {
 
-    this.clienteService.deleteCliente(idCliente).subscribe({
-      next: (data) => {
-        console.log(data);
-        this.getClientes();
-      },
-      error: (error) => {
-        console.log(error);
+    Swal.fire({
+      title: '¿Estas seguro de eliminar el cliente?',
+      text: "¡No podrás revertir esto!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33'
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.clienteService.deleteCliente(idCliente).subscribe({
+          next: (data) => {
+            console.log(data);
+            this.getClientes();
+          },
+          error: (error) => {
+            console.log(error);
+          }
+        })
       }
     })
-
   }
 
 }
